fixed the QuabLog's memory leak by using basedebug
