+++ src/Quab/debug/QuabLog.h
@@ -2,15 +2,14 @@
 #define QUAB_DEBUG_LOG_H
 
 #include "../QuabDef.h"
-#define QUAB_DEFAULT_LOG_DIR "./log/"
 
 namespace Quab
 {
     class QUAB_API QuabLog
 	{
 	public:		
-		static void Log(const char *msg);		
-		static void Error(const char *msg);
+		static void Log(const char *moduleName, const char *msg);		
+		static void Error(const char *moduleName, const char *msg);
 	};	
 }
