+++ src/Quab/qbase/basedebug.c
@@ -5,14 +5,61 @@
 #include <time.h>
 #include <stdlib.h>
 
-#ifdef QBASE_OS_WIN32
+#ifdef QUAB_OS_WIN32
 #include <direct.h>
+#include <windows.h>
 #define LOG_PATH "log\\"
 #else
 #include <sys/stat.h>
 #include <sys/types.h>
-#define LOG_PATH "log/"
+#define LOG_PATH "./log/"
 #endif
+
+enum PRINT_COLOR{
+	COLOR_GREEN  = 10,
+	COLOR_BLUE   = 11,
+	COLOR_RED    = 12,
+	COLOR_YELLOW = 14
+};
+
+/*    color printing helper    */
+static void
+_printColorText(const char *text, int color) {
+    if(color <= 0)  {
+        printf(text);
+    }
+    else    {
+#ifdef QUAB_OS_WIN32
+        WORD colorOld;
+        HANDLE handle = GetStdHandle(STD_OUTPUT_HANDLE);
+        CONSOLE_SCREEN_BUFFER_INFO csbi;
+        GetConsoleScreenBufferInfo(handle, &csbi);
+        colorOld = csbi.wAttributes;
+        SetConsoleTextAttribute(handle, (WORD)color);
+        printf(text);
+        SetConsoleTextAttribute(handle, colorOld);
+#else
+        switch(color)
+        {
+        case COLOR_BLUE:
+            printf("\033[34m%s\033[0m", text);
+            break;
+        case COLOR_GREEN:
+            printf("\033[32m%s\033[0m", text);
+            break;
+        case COLOR_RED:
+            printf("\033[31m%s\033[0m", text);
+            break;
+        case COLOR_YELLOW:
+            printf("\033[33m%s\033[0m", text);
+            break;
+        default:
+            return;
+        }
+#endif
+    }
+}
+
 
 static char *
 debug_timestring()	{
@@ -56,6 +103,7 @@ debug_filestring()	{
 static void
 debug_writemsg(const char * status, const char *obj, const char *msg)	{
 	int len = 0;
+	int color = -1;
 	char *filename = debug_filestring();
 	char *tstr = debug_timestring();
 	char *logfile = NULL;
@@ -63,7 +111,7 @@ debug_writemsg(const char * status, const char *obj, const char *msg)	{
 
 	len += strlen(LOG_PATH);
 	len += strlen(filename);
-#ifdef QBASE_OS_LINUX
+#ifdef QUAB_OS_LINUX
 	mkdir(LOG_PATH, S_IRWXU);
 #else
 	_mkdir(LOG_PATH);
@@ -79,7 +127,21 @@ debug_writemsg(const char * status, const char *obj, const char *msg)	{
 		fclose(f);
 	}
 	free(logfile);
-	printf("[%s]<%s> IN \"%s\" :%s\n", tstr, status != NULL? status: "MESSAGE", obj, msg);
+//	printf("[%s]<%s> IN \"%s\" :%s\n", tstr, status != NULL? status: "MESSAGE", obj, msg);
+#ifdef QUAB_STATUS_DEBUG
+	printf("[%s]",tstr);
+	status = status != NULL? status : "MESSAGE";
+	if(strcmp(status, "MESSAGE") == 0)
+		color = COLOR_BLUE;
+	else if(strcmp(status, "ERROR") == 0)
+		color = COLOR_RED;
+	else if(strcmp(status, "DEBUG") == 0)
+		color = COLOR_YELLOW;
+	_printColorText("<",color);
+	_printColorText(status, color);
+	_printColorText(">",color);
+	printf(" IN \"%s\" :%s\n", obj, msg);
+#endif
 	free(tstr);
 }
